Fixes some warnings throughout and fixes an issue where heat flux was not being re-dimensionalized for history file output.

+++ Common/src/graph_coloring_structure.cpp
@@ -241,7 +241,7 @@ void CGraphColoringStructure::GraphVertexColoring(
       colorLocalVertices.resize(nLocalVert);
 
       for(unsigned long i=0; i<nLocalVert; ++i)
-        colorLocalVertices[i] = i + nVerticesPerRank[myRank];
+        colorLocalVertices[i] = (int)(i + nVerticesPerRank[myRank]);
 
       break;
     }
